Re-organise downloads related files
