+++ Model/Downloads/DownloadService.swift
@@ -13,105 +13,10 @@
 // You should have received a copy of the GNU General Public License
 // along with Kiwix; If not, see https://www.gnu.org/licenses/.
 
-//
-//  DownloadService.swift
-//  Kiwix
-
 import Combine
 import CoreData
 import UserNotifications
 import os
-
-struct DownloadState: Codable {
-    let downloaded: Int64
-    let total: Int64
-    let resumeData: Data?
-    
-    var isPaused: Bool {
-        resumeData != nil
-    }
-
-    static func empty() -> DownloadState {
-        .init(downloaded: 0, total: 1, resumeData: nil)
-    }
-
-    init(downloaded: Int64, total: Int64, resumeData: Data?) {
-        guard total >= downloaded, total > 0 else {
-            assertionFailure("invalid download progress values: downloaded \(downloaded) total: \(total)")
-            self.downloaded = downloaded
-            self.total = downloaded
-            self.resumeData = resumeData
-            return
-        }
-        self.downloaded = downloaded
-        self.total = total
-        self.resumeData = resumeData
-    }
-
-    func updatedWith(downloaded: Int64, total: Int64) -> DownloadState {
-        DownloadState(downloaded: downloaded, total: total, resumeData: resumeData)
-    }
-
-    func updatedWith(resumeData: Data?) -> DownloadState {
-        DownloadState(downloaded: downloaded, total: total, resumeData: resumeData)
-    }
-}
-
-@MainActor
-final class DownloadTasksPublisher {
-
-    let publisher: CurrentValueSubject<[UUID: DownloadState], Never>
-    private var states = [UUID: DownloadState]()
-
-    init() {
-        publisher = CurrentValueSubject(states)
-        if let jsonData = UserDefaults.standard.object(forKey: "downloadStates") as? Data,
-           let storedStates = try? JSONDecoder().decode([UUID: DownloadState].self, from: jsonData) {
-            states = storedStates
-            publisher.send(states)
-        }
-    }
-
-    func updateFor(uuid: UUID, downloaded: Int64, total: Int64) {
-        if let state = states[uuid] {
-            states[uuid] = state.updatedWith(downloaded: downloaded, total: total)
-        } else {
-            states[uuid] = DownloadState(downloaded: downloaded, total: total, resumeData: nil)
-        }
-        publisher.send(states)
-        saveState()
-    }
-
-    func resetFor(uuid: UUID) {
-        states.removeValue(forKey: uuid)
-        publisher.send(states)
-        saveState()
-    }
-
-    func isEmpty() -> Bool {
-        states.isEmpty
-    }
-
-    func resumeDataFor(uuid: UUID) -> Data? {
-        states[uuid]?.resumeData
-    }
-
-    func updateFor(uuid: UUID, withResumeData resumeData: Data?) {
-        if let state = states[uuid] {
-            states[uuid] = state.updatedWith(resumeData: resumeData)
-            publisher.send(states)
-            saveState()
-        } else {
-            assertionFailure("there should be a download task for: \(uuid)")
-        }
-    }
-    
-    private func saveState() {
-        if let jsonStates = try? JSONEncoder().encode(states) {
-            UserDefaults.standard.setValue(jsonStates, forKey: "downloadStates")
-        }
-    }
-}
 
 final class DownloadService: NSObject, URLSessionDelegate, URLSessionTaskDelegate, URLSessionDownloadDelegate {
     static let shared = DownloadService()
